TikzTable Creator
The brain of the project, ironically being a batch file. It's actually the first programming language I ever learnt so it's funny it plays such an important role in the project.
This file basically calls every python file to generate the text files, generates the csv files, and deletes everything that's useless at the end. Except for the extension-less files null and %urlval%, sorry about those I really don't know what to do about them.